fetch: Do not trim the end of the displayed url
Ever since fetch was ported to C, the end of the url printed to the
screen has been truncated to remove the trailing "/" and the trailing
".git". As far as I can tell this is the only place in git that does
such a thing.

This patch adds the "/" and ".git" back to the end of url by removing the
code that would normally truncate it. This caches the length of the url
so that it is no longer recalculated for each ref.

Output before this patch:

	$ git fetch remote-with-dot-git-and-slash
	From https://github.com/git/git
	   a155a5f..5512ac5  maint      -> upstream/maint

Output after this patch:

	$ git fetch remote-with-dot-git-and-slash
	From https://github.com/git/git.git/
	   a155a5f..5512ac5  maint      -> upstream/maint